Remove version restriction on byebug.
Latest version of byebug is 3.5.1, so let’s not restrict byebug at all. The `byebug` method seems to be a fairly stable interface to byebug starting the debugger in code crossing versions.

The reason to restrict minitest is because of the plug-in interface; if the interface changes in an incompatible way in minitest 6 (not likely, but possible), then we don’t want the minitest interface to break. The single interface to byebug seems to be safe to trust for any future version.